Separate expanded and collapsed jsx, add animation and styles
